ElasticsearchIndexer will attempt to use rendition service to extract text from images
Fixes 80

+++ spring-content-elasticsearch/src/test/java/org/springframework/content/elasticsearch/ElasticsearchIT.java
@@ -1,37 +1,12 @@
 package org.springframework.content.elasticsearch;
 
-import java.time.Duration;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-import com.github.paulcwarren.ginkgo4j.Ginkgo4jSpringRunner;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.elasticsearch.action.get.GetRequest;
-import org.elasticsearch.action.get.GetResponse;
-import org.elasticsearch.client.RequestOptions;
-import org.elasticsearch.client.RestHighLevelClient;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import pl.allegro.tech.embeddedelasticsearch.EmbeddedElastic;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.content.commons.annotations.ContentId;
-import org.springframework.content.commons.repository.ContentStore;
-import org.springframework.content.commons.search.Searchable;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.test.context.ContextConfiguration;
-
 import static com.github.grantwest.eventually.EventuallyLambdaMatcher.eventuallyEval;
 import static com.github.paulcwarren.ginkgo4j.Ginkgo4jDSL.AfterEach;
 import static com.github.paulcwarren.ginkgo4j.Ginkgo4jDSL.BeforeEach;
 import static com.github.paulcwarren.ginkgo4j.Ginkgo4jDSL.Context;
 import static com.github.paulcwarren.ginkgo4j.Ginkgo4jDSL.Describe;
 import static com.github.paulcwarren.ginkgo4j.Ginkgo4jDSL.It;
+import static com.github.paulcwarren.ginkgo4j.Ginkgo4jDSL.FIt;
 import static org.hamcrest.CoreMatchers.allOf;
 import static org.hamcrest.CoreMatchers.hasItem;
 import static org.hamcrest.CoreMatchers.hasItems;
@@ -39,19 +14,46 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
 
+import java.time.Duration;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+import org.elasticsearch.action.get.GetRequest;
+import org.elasticsearch.action.get.GetResponse;
+import org.elasticsearch.client.RequestOptions;
+import org.elasticsearch.client.RestHighLevelClient;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.content.commons.annotations.ContentId;
+import org.springframework.content.commons.annotations.MimeType;
+import org.springframework.content.commons.renditions.Renderable;
+import org.springframework.content.commons.repository.ContentStore;
+import org.springframework.content.commons.search.Searchable;
+import org.springframework.content.elasticsearch.ElasticsearchIT.Document;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.test.context.ContextConfiguration;
+
+import com.github.paulcwarren.ginkgo4j.Ginkgo4jConfiguration;
+import com.github.paulcwarren.ginkgo4j.Ginkgo4jSpringRunner;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
 @RunWith(Ginkgo4jSpringRunner.class)
 //@Ginkgo4jConfiguration(threads=1)
 @ContextConfiguration(classes = {EmbeddedElasticConfig.class, ElasticsearchConfig.class})
-public class ElasticsearchIntegrationTest {
+public class ElasticsearchIT {
 
 	@Autowired
 	private DocumentRepository repo;
 
 	@Autowired
 	private DocumentContentStore store;
-
-	@Autowired
-	private EmbeddedElastic server;
 
 	@Autowired
 	private RestHighLevelClient client;
@@ -79,11 +81,15 @@ public class ElasticsearchIntegrationTest {
 				});
 
 				AfterEach(() -> {
-					store.unsetContent(doc1);
-					repo.delete(doc1);
+				    if (doc1 != null) {
+    					store.unsetContent(doc1);
+    					repo.delete(doc1);
+				    }
 
-					store.unsetContent(doc2);
-					repo.delete(doc2);
+				    if (doc2 != null) {
+    					store.unsetContent(doc2);
+    					repo.delete(doc2);
+				    }
 				});
 
 				It("should index the documents", () -> {
@@ -123,6 +129,24 @@ public class ElasticsearchIntegrationTest {
 								Duration.ofSeconds(10)));
 					});
 				});
+				
+	            Context("given a text extracting renderer", () -> {
+
+                    BeforeEach(() -> {
+                        doc1 = new Document();
+                        doc1.setTitle("doc 1");
+                        doc1.setAuthor("author@email.com");
+                        doc1.setMimeType("image/png");
+                        doc1 = store.setContent(doc1, this.getClass().getResourceAsStream("/image.png"));
+                        doc1 = repo.save(doc1);
+                    });
+
+                    It("should index the documents", () -> {
+                        assertThat(() -> store.search("wisdom"), eventuallyEval(
+                                    hasItem(doc1.getContentId()),
+                                Duration.ofSeconds(10)));
+                    });
+	            });
 
 				Context("given that document is deleted", () -> {
 
@@ -134,6 +158,11 @@ public class ElasticsearchIntegrationTest {
 						id2 = doc2.getContentId();
 						store.unsetContent(doc2);
 						repo.delete(doc2);
+					});
+					
+					AfterEach(() -> {
+                       doc1 = null;
+                       doc2 = null;
 					});
 
 					It("should delete the record of the content from the index", () -> {
@@ -158,7 +187,7 @@ public class ElasticsearchIntegrationTest {
 		//
 	}
 
-	public interface DocumentContentStore extends ContentStore<Document, String>, Searchable<String> {
+	public interface DocumentContentStore extends ContentStore<Document, String>, Searchable<String>, Renderable<Document> {
 		//
 	}
 
@@ -174,6 +203,9 @@ public class ElasticsearchIntegrationTest {
 
 		@ContentId
 		private String contentId;
+		
+		@MimeType
+		private String mimeType;
 
 		private String title;
 		private String author;
